ajustado visualização de compras e separação do componentes
